requirements.txt
alpaca-py
alpaca-trade-api
numpy
pandas
pyarrow
scipy
loguru
decouple
matplotlib
seaborn
tabulate
streamlit
plotly
python-dotenv
